backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
stripe==7.0.0

# For development
black==23.9.1
flake8==6.1.0
pytest==7.4.2
pytest-django==4.5.2